Use packagized header and footer
