--- client/layout/universal-navbar-header/btn-menu-item.jsx
@@ -1,9 +1,0 @@
-const UniversalNavbarBtnMenuItem = ( { elementContent, className } ) => {
-	return (
-		<button role="menuitem" className={ className }>
-			{ elementContent } <span className="x-nav-link-chevron" aria-hidden="true"></span>
-		</button>
-	);
-};
-
-export default UniversalNavbarBtnMenuItem;
